+++ dm-20151123/alibabacloud_dm20151123/client.py
@@ -913,6 +913,88 @@ class Client(OpenApiClient):
         runtime = util_models.RuntimeOptions()
         return await self.delete_domain_with_options_async(request, runtime)
 
+    def delete_invalid_address_with_options(
+        self,
+        request: dm_20151123_models.DeleteInvalidAddressRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> dm_20151123_models.DeleteInvalidAddressResponse:
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
+        if not UtilClient.is_unset(request.resource_owner_id):
+            query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.to_address):
+            query['ToAddress'] = request.to_address
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='DeleteInvalidAddress',
+            version='2015-11-23',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            dm_20151123_models.DeleteInvalidAddressResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def delete_invalid_address_with_options_async(
+        self,
+        request: dm_20151123_models.DeleteInvalidAddressRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> dm_20151123_models.DeleteInvalidAddressResponse:
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
+        if not UtilClient.is_unset(request.resource_owner_id):
+            query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.to_address):
+            query['ToAddress'] = request.to_address
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='DeleteInvalidAddress',
+            version='2015-11-23',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            dm_20151123_models.DeleteInvalidAddressResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def delete_invalid_address(
+        self,
+        request: dm_20151123_models.DeleteInvalidAddressRequest,
+    ) -> dm_20151123_models.DeleteInvalidAddressResponse:
+        runtime = util_models.RuntimeOptions()
+        return self.delete_invalid_address_with_options(request, runtime)
+
+    async def delete_invalid_address_async(
+        self,
+        request: dm_20151123_models.DeleteInvalidAddressRequest,
+    ) -> dm_20151123_models.DeleteInvalidAddressResponse:
+        runtime = util_models.RuntimeOptions()
+        return await self.delete_invalid_address_with_options_async(request, runtime)
+
     def delete_ipfilter_by_edm_id_with_options(
         self,
         request: dm_20151123_models.DeleteIpfilterByEdmIdRequest,
@@ -1974,8 +2056,12 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.domain_name):
             query['DomainName'] = request.domain_name
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
         if not UtilClient.is_unset(request.password):
             query['Password'] = request.password
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
         if not UtilClient.is_unset(request.resource_owner_id):
             query['ResourceOwnerId'] = request.resource_owner_id
         req = open_api_models.OpenApiRequest(
@@ -2006,8 +2092,12 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.domain_name):
             query['DomainName'] = request.domain_name
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
         if not UtilClient.is_unset(request.password):
             query['Password'] = request.password
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
         if not UtilClient.is_unset(request.resource_owner_id):
             query['ResourceOwnerId'] = request.resource_owner_id
         req = open_api_models.OpenApiRequest(